Feature/extension functionality done (#53)
* FIXED arrangeDone twice error

* UI Update (#36)

* EDITED qml path

* FIXED Uppertab button image bug

* ADDED Upperbutton hover

* EDITED fileOpen Threading, UI threading

* ADDED LeftPartList adding & State change done

* EDITED cut signals, variables, qm not found error

* EDITED cut mode selection

* ADDED Add to Part List

* ADDED Part List Visible Done

* ADDED Part List Delete

* ADDED Part List Name Parsing Done

* EDITED leftmodel rightmodel to be child, model cut selection working on

* MERGED develop branch

* bug fix

* ADDED mesh repair threading

* - add mesh rotate to Mesh class
- add rotatemodelmesh to GLModel class
- initilize the model position when load stl file

* FIXED glmodel tosparse twice / file reopen slowdown

* EDITED rotatemesh sparsemesh calc mechanism

* ADDED vertexRotate code on auto arrangement

* - Rotation function workingproto
- todo : Direction of rotation must be change with orientation of camera
- todo : multiple model rotation (selected one)

* splash change

* ADDED vertexRotate code to autoorientation

* ADDED Add mainwindow address in GLModel

* FIXED merge bug fix

* EDITED plate done (#31)

* ADDED stlexporter, slicing engine part

* ADDED DentEngine repository

* - move prototype dev
- rotate & move size fix

* - move function developt

* - add application icon

* - move & rotation optimization

* ADDED export feature, slicing feature

* ROLLBACK uppertab.qml

* - move direction fix

* FIXED toSparse Error

* - coordinate system change (view rotae&move)
- change world2screen function

* FIXED shared component error, added selectedmodel, added pick signal to glmodel

* EDITED UpperButton Font Size & Label pop UI

* ADDED View mode function done

* EDITED Move Arrow Color

* EDITED Rotate Torus Color

* - rotate direction fix

* -mesh scale function

* Bug Fixed (#37)

* EDITED qml path

* FIXED Uppertab button image bug

* ADDED Upperbutton hover

* EDITED fileOpen Threading, UI threading

* ADDED LeftPartList adding & State change done

* EDITED cut signals, variables, qm not found error

* EDITED cut mode selection

* ADDED Add to Part List

* ADDED Part List Visible Done

* ADDED Part List Delete

* ADDED Part List Name Parsing Done

* EDITED leftmodel rightmodel to be child, model cut selection working on

* MERGED develop branch

* bug fix

* ADDED mesh repair threading

* - add mesh rotate to Mesh class
- add rotatemodelmesh to GLModel class
- initilize the model position when load stl file

* FIXED glmodel tosparse twice / file reopen slowdown

* EDITED rotatemesh sparsemesh calc mechanism

* ADDED vertexRotate code on auto arrangement

* - Rotation function workingproto
- todo : Direction of rotation must be change with orientation of camera
- todo : multiple model rotation (selected one)

* splash change

* ADDED vertexRotate code to autoorientation

* ADDED Add mainwindow address in GLModel

* FIXED merge bug fix

* EDITED plate done (#31)

* ADDED stlexporter, slicing engine part

* ADDED DentEngine repository

* - move prototype dev
- rotate & move size fix

* - move function developt

* - add application icon

* - move & rotation optimization

* ADDED export feature, slicing feature

* ROLLBACK uppertab.qml

* - move direction fix

* FIXED toSparse Error

* - coordinate system change (view rotae&move)
- change world2screen function

* FIXED shared component error, added selectedmodel, added pick signal to glmodel

* EDITED UpperButton Font Size & Label pop UI

* ADDED View mode function done

* EDITED Move Arrow Color

* EDITED Rotate Torus Color

* FIXED PartList add part bug

* ADDED Initial Camera Position

* ADDED Model Delete function

* “forvideo”

* FIXING object picker by scene3d -> canvas3d

* FIXING mesh bugs

* ADDED DentEngine

* INTEGRATED DentEngine

* EDITED DentEngine version

* ERROR labellingTextPreview on glmodel creation

* EDITED DentEngine version

* EDITED slicing engine execution to be done background

* ADDED poly2tri, clip2tri libraries

* MOVED polyclipping to DentEngine

* EDITED toSparse function to be scalable

* FIXED logarithm scale value

* FIXED mesh setcount 12 error

* ADDED shelloffset function and done 70%

* MOVED meshvertex2hash, intpoint2hash function to mesh.cpp

* ADDED model selection connectHandler

* EDITED extension function

* EDITING detectingoutline

* working on extension, edited whitening problem after labelling

* DELETE unused debug texts

* EDITED model cut function call createModelFile from glmodel

* FIXED model move error on moving&select canceled

* works fine intermitantly

* FIXED shell offset errors

* UPDATE DentEngine

* FIXED selection error occured on wrong merge

* FIXED slicing engine not to save & export

* 승환 fix

* FIXED model selectino

* DONE extension

+++ glmodel.h
@@ -232,7 +232,7 @@ public slots:
     void closeExtension();
     void colorExtensionFaces();
     void uncolorExtensionFaces();
-    void generateExtensionFaces(float distance);
+    void generateExtensionFaces(double distance);
 
     // ShellOffset
     void generateShellOffset(double factor);
